+++ Collection Manager/Plugin.cs
@@ -1,14 +1,14 @@
 using System.Globalization;
-using Jellyfin.Plugin.MediaCollectionManager.Configuration;
-using Jellyfin.Plugin.MediaCollectionManager.Models;
+using Jellyfin.Plugin.CollectionManager.Configuration;
+using Jellyfin.Plugin.CollectionManager.Models;
 using MediaBrowser.Common.Configuration;
 using MediaBrowser.Common.Plugins;
 using MediaBrowser.Model.Plugins;
 using MediaBrowser.Model.Serialization;
 
-namespace Jellyfin.Plugin.MediaCollectionManager;
+namespace Jellyfin.Plugin.CollectionManager;
 
-/// <summary>The Media Collection Manager plugin entry point.</summary>
+/// <summary>The Collection Manager plugin entry point.</summary>
 public sealed class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
 {
     private readonly object _configurationLock = new();
@@ -23,7 +23,7 @@ public sealed class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
     public static Plugin? Instance { get; private set; }
 
     /// <inheritdoc />
-    public override string Name => "Media Collection Manager";
+    public override string Name => "Collection Manager";
 
     /// <inheritdoc />
     public override Guid Id => Guid.Parse("5e9d033a-41dd-4c93-b53f-aa94f8e2f7e9");
@@ -34,7 +34,7 @@ public sealed class Plugin : BasePlugin<PluginConfiguration>, IHasWebPages
         yield return new PluginPageInfo
         {
             Name = Name,
-            DisplayName = "Media Collection Manager",
+            DisplayName = "Collection Manager",
             EnableInMainMenu = true,
             EmbeddedResourcePath = string.Format(
                 CultureInfo.InvariantCulture,